refactored App.js\nconverted indentation to spaces\nset indentation to 4 spaces\nrefactored file names across project\nconverted App.js to React class\nmoved results table into App scope
