--- src/components/results-table/usePagination.js
@@ -1,30 +1,0 @@
-import { useState } from "react";
-import buildQueryString from "../../utils/query-url"
-import { api, API_RESULT_KEYS } from "../../http/apiClient";
-
-export default function usePagination(searchResult) {
-    // State helpers
-    const [page, setPage] = useState(0);
-    const [rowsPerPage, setRowsPerPage] = useState(10);
-
-  // Callback for changing a page
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  // Callback for managing new rows per page
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(parseInt(event.target.value));
-    setPage(0);
-  };
-
-  const emptyRows = rowsPerPage - Math.min(rowsPerPage, searchResult.map((result) => result.length) - page * rowsPerPage);
-   
-    return {
-        page,
-        rowsPerPage,
-        handleChangePage,
-        handleChangeRowsPerPage,
-        emptyRows
-    }
-}
