+++ src/components/resultstable/Table.css
@@ -1,0 +1,16 @@
+.table {
+  margin-top: 80px;
+  margin-bottom: 100px;
+  width: 100%;
+}
+
+table {
+  width: 100%;
+  font-size: 1rem;
+}
+
+th,
+td {
+  padding: 10px;
+  box-shadow: 1px 1px rgb(194, 194, 194);
+}
